test(routing): restore the registered canonical ULW seam
Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

Constraint: The registered SC2 command targets tests.test_router_content.CanonicalUlwWorkflowNameTests.

Rejected: Keep the class in a dedicated module | the required dotted test binding would remain broken

Confidence: high

Scope-risk: narrow

Directive: Keep the registered dotted test path stable.

Tested: exact SC2 test; canonical ULW 3 tests; router content 108 tests; full 8675 tests; compileall; Ruff; LSP; diff-check

Not-tested: GitHub CI for this follow-up head
